more work new classifier dataset
I think this might actually be what we want now?
although I think one issues is that since the state history from the
origin is not included in the examples, some examples are undecidable

I could also try including the "true" state from the origin, but that
would also not match our problem setup. The better way is to include the
initial tether state in the classifier image.