tested obstacle skirting with higher threshold (looser) for "close"
now obstacle skirting actually shows converging. before it rarely did.
you had to pull for a lot of time steps. This is partially why we are
getting so few examples. Now is it ok to increase this threshold? maybe
not, because it will make our plans less accurate. But you could see it
as a tradeoff between accuracy and "completeness" (how often does it
find a path when one exists)